Injects a synthetic sample to track overhead.
This version creates a synthetic sample called 'overhead'
to track walltime so it's easier to look at a flamegraph
and see that it's bound by something other than CPU.

It can be used as a filter to rewrite a series of profiles.
Running it like

driver sample.pb.gz sample.pb.gz

will update the samples in place. Useful for rewriting a collection
of profiles.